fixing code review changes in fdt.c and fdt.h

+++ kernel/fdt.h
@@ -27,10 +27,11 @@ uint32_t fdt32_to_cpu(uint32_t value);
 // converts FDT's 64 bit big endian to CPU's native little endian:
 uint64_t fdt64_to_cpu(uint64_t value);
 
-//uses the fdt_address from a1 reg and checks the magic field for 0xDDFEED
+//uses the fdt_address from a1 reg and checks the magic field for 0xD00DFEED
 int check_fdt_magic(uintptr_t fdt_address);
 
-//fn reads the string value stored at (base + offset) address. 
-void read_string(uint64_t base, uint32_t offset);
+//reads the string value stored at (base + offset) address. 
+void read_string(uintptr_t base, uint32_t offset);
+
 #endif
